Rename karton.py to program.py
Initially it was called main.py, then, being the main entry point of the
program, it was renamed to karton.py.
Now it's not the main entry point (as we have scripts/karton) and it's
not even an executable any more, so it's better to give it a more
descriptive name.