dockerctl: use 'podman' if present and 'docker version' fails
`podman` is an almost-CLI-compatible replacement for `docker`, which
allows unprivileged users to run containers, with no daemon. On some
systems, `docker` is a symbolic link to `podman`.

I say "almost" because one command that is not compatible is `docker
version --format '{{ json . }}'`, used by Karton to detect whether
docker is present and usable.
https://github.com/containers/libpod/issues/2671

To work around this, if _try_docker() fails, but running `podman
version` succeeds, use `podman` instead. Otherwise, behave as before.