WIP: support using UID 0 inside the container
When using podman as an unprivileged user, UID 0 in the container is
mapped to your real UID outside the container. Karton's behaviour of
creating a "mirror" unprivileged user inside the container is
undesirable because its UID inside the container (in my case, 1001) is
mapped to something else (in my case, 166536) outside, so I now have
files in my homedir that I don't own.

Previously, setting props.uid = 0 would fail with:

    STEP 26: RUN     mkdir -p $(dirname /sysroot/home/wjt) &&
                     useradd -m -s /bin/bash --home-dir /sysroot/home/wjt --uid 0 wjt &&
                     chown wjt /sysroot/home/wjt
    useradd: UID 0 is not unique

With this change, setting:

    props.uid = 0
    props.sudo = props.SUDO_NO

seems to do the right thing.

This commit is WIP because:

* props.username is part-ignored with this change,
* I think this behaviour should be used automatically when using podman
  as an unprivileged user
* I haven't tested it much